Structs in structs complete

+++ output/output.c
@@ -260,7 +260,7 @@ print_string("beta.g.a = ");
 set_pointer_i32((i32*)(&((beta_2.g).a)),9000);
 println_i32((i32)((beta_2.g).a));
 print_string("beta.g.a = ");
-println_i32((i32)beta_2.g.a);
+println_i32(beta_2.g.a);
 }
 }
 void files_0()
